# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/raw/create": {
            "post": {
                "tags": [
                    "Raw Stock"
                ],
                "summary": "Stock Insert",
                "description": "Raw Material Insertion",
                "operationId": "Insert",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "raw_name",
                                    "unit",
                                    "stock",
                                    "price"
                                ],
                                "properties": {
                                    "raw_name": {
                                        "type": "text",
                                        "example": "test"
                                    },
                                    "unit": {
                                        "type": "number",
                                        "example": "10"
                                    },
                                    "stock": {
                                        "type": "text",
                                        "example": "test"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": "1000"
                                    },
                                    "user_id": {
                                        "type": "number",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/raw/update": {
            "patch": {
                "tags": [
                    "Raw Stock"
                ],
                "summary": "Stock Update",
                "description": "Raw Material Update",
                "operationId": "Update",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "raw_name",
                                    "unit",
                                    "stock",
                                    "price",
                                    "user_id",
                                    "raw_id"
                                ],
                                "properties": {
                                    "raw_id": {
                                        "type": "intiger",
                                        "example": "3"
                                    },
                                    "raw_name": {
                                        "type": "text",
                                        "example": "test"
                                    },
                                    "unit": {
                                        "type": "number",
                                        "example": "10"
                                    },
                                    "stock": {
                                        "type": "text",
                                        "example": "test"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": "1000"
                                    },
                                    "user_id": {
                                        "type": "number",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/raw/details/{id}": {
            "get": {
                "tags": [
                    "Raw Stock"
                ],
                "summary": "Get list of Raw detail",
                "description": "Returns list of packages detail",
                "operationId": "getPackageDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/raw/all": {
            "get": {
                "tags": [
                    "Raw Stock"
                ],
                "summary": "Stock list ",
                "description": "Stock list",
                "operationId": "fetch all",
                "parameters": [
                    {
                        "name": "search_text",
                        "in": "path",
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "start",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "length",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order",
                        "in": "path",
                        "schema": {
                            "type": "text"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/raw/delete/{id}": {
            "delete": {
                "tags": [
                    "Raw Stock"
                ],
                "summary": "Delete ",
                "description": "delete",
                "operationId": "Delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "number",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "bc420"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "UnAuthorised"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "roles",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "example": "sharvari"
                                    },
                                    "email": {
                                        "type": "text",
                                        "example": "sharvari@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "123456789"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "example": "123456789"
                                    },
                                    "roles": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get the details of an authenticated user",
                "description": "",
                "operationId": "getAuthUser",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "schema": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error xXx"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}